Add deep sleep example to wakeOnMove
The LIS3DH interrupt appears to still wake the electron out of deep sleep mode without the WKP and RISING variables passed to System.sleep(). 

I found this interesting when I was testing the code, and thought it was worth sharing. I have the coin cell battery and lipo attached to 
the asset tracker Not sure if it makes a difference to the functionality.

Thanks for making the library. I appreciate it.